requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
openai>=1.0
aiohttp>=3.9.0
edge-tts>=6.1.12
opuslib>=3.0.1
pydantic>=2.0.0
yt-dlp>=2025.1.0
authlib>=1.3.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
itsdangerous>=2.0.0
python-multipart>=0.0.6
Pillow>=10.0.0

